Add native FP16 and logical-position causal SM90 micro-prefill
Validate both retained families on Lightning and Modal H100 against FP32 output/LSE references, including mutable graphs and empty visibility. Preserve the mixed deferred-denominator ablation, compiler research, and raw evidence without promoting a new performance route.

+++ tests/test_sm90_micro_prefill.py
@@ -82,8 +82,8 @@ def test_shape_contract_spans_m2_to_m64_without_cell_registration():
 
 
 def test_shape_contract_reports_semantic_boundaries():
-    q = torch.empty(1, 1, 12, 96, dtype=torch.float16, device="meta")
-    kv = torch.empty(1, 2, 4097, 96, dtype=torch.float16, device="meta")
+    q = torch.empty(1, 1, 12, 96, dtype=torch.float32, device="meta")
+    kv = torch.empty(1, 2, 4097, 96, dtype=torch.float32, device="meta")
     reasons = micro_prefill_shape_reasons(q, kv, kv)
     assert set(reasons) == {"query_len", "gqa", "head_dim", "kv_len", "dtype"}
